// src/generated/model/userAdminResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { RoleInfo } from './roleInfo.ts';

export interface UserAdminResponse {
  id?: string;
  fullName?: string;
  username?: string;
  email?: string;
  telephone?: string;
  accessLevel?: string;
  status?: string;
  accountLocked?: boolean;
  lockReason?: string;
  lockedAt?: string;
  requirePasswordChange?: boolean;
  tenantId?: string;
  tenantName?: string;
  roles?: RoleInfo[];
  lastLoginAt?: string;
  createdAt?: string;
  updatedAt?: string;
  version?: number;
}